+ fixed nasty bug in gfsm_alphabet_load_handle() and gfsmio_getc() causing char decimal 255 to return as EOF
git-svn-id: svn+ssh://svn.dwds.de/home/moocow/svn/public/gfsm/trunk@264 097f6656-c427-11dc-a790-1b5d5187385a

+++ src/libgfsm/gfsmIO.h
@@ -38,7 +38,7 @@
 #include <unistd.h>
 #include <stdio.h>
 
-#define GFSMIO_EOF -1
+#define GFSMIO_EOF ((int)-1)
 
 /*======================================================================
  * I/O: types
